Get public keys from keystore
