+++ app/build.gradle
@@ -6,7 +6,7 @@ android {
     compileSdk 31
 
     defaultConfig {
-        applicationId "com.test.eosio"
+        applicationId "com.test.ekis"
         minSdk 22
         targetSdk 31
         versionCode 1
