+++ app/src/main/AndroidManifest.xml
@@ -8,7 +8,7 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.Eosio">
+        android:theme="@style/Theme.ekis">
         <activity
             android:name="com.test.ekisdemo.MainActivity"
             android:exported="true">
