+++ ekisandroidkeystoresignatureprovider/package-info.md
@@ -1,12 +1,12 @@
-# Package one.block.eosiojavaandroidkeystoresignatureprovider
+# Package one.block.ekisjavaandroidkeystoresignatureprovider
 
-## Main package of EOSIO Signature Provider for Android KeyStore
+## Main package of ekis Signature Provider for Android KeyStore
 
 It includes the signature provider and utility class, which provides many convenient methods for key management (e.g. key generation, query, deletion...).
 
-# Package one.block.eosiojavaandroidkeystoresignatureprovider.errors
+# Package one.block.ekisjavaandroidkeystoresignatureprovider.errors
 
-## Error package of EOSIO Signature Provider for Android KeyStore
+## Error package of ekis Signature Provider for Android KeyStore
 
 It includes exceptions and error message definitions that are used for error handling throughout the Android KeyStore Signature Provider library.
